init berth_id assign

+++ src/Berth.cpp
@@ -11,7 +11,7 @@ Berth::Berth(int x, int y, int transport_time, int loading_speed, int id_berth) 
     this->area = Area(x, x+4, y, y+4);
     this->transport_time = transport_time;
     this->loading_speed = loading_speed;
-    this->value = loading_speed/transport_time;
+    this->value = (double)loading_speed / transport_time;
     this->num_assigned_robots = 0;
     for (int i = 0; i < 4; i++){
         for (int j = 0; j < 4; j++){
